Background, Content_Script, Manifest and Options
1. Added background.js. 
I've put in a checking to check if the the version is the same as the version that we are using. If not, we will launch the options.html. 

2. Added content_script.js
I've put in a text selection function that will return the selected text. 

3. Manifest.json
Added background to cater for background logic. 
Added content_scripts to enable extension to read details of web pages. 
Added content_script.js in to the list. It will inject this script into the html. 
Added incognito into it. I used split mode since we are not making any calls to server. I should change it to another mode if I were to make calls to the server. 
Added options_page and options_ui for our options. 
Added permissions of alarm and notification. 

4. Added options.html page. I will write this later.