Closing of #1 (#8)
* Background, Content_Script, Manifest and Options

1. Added background.js. 
I've put in a checking to check if the the version is the same as the version that we are using. If not, we will launch the options.html. 

2. Added content_script.js
I've put in a text selection function that will return the selected text. 

3. Manifest.json
Added background to cater for background logic. 
Added content_scripts to enable extension to read details of web pages. 
Added content_script.js in to the list. It will inject this script into the html. 
Added incognito into it. I used split mode since we are not making any calls to server. I should change it to another mode if I were to make calls to the server. 
Added options_page and options_ui for our options. 
Added permissions of alarm and notification. 

4. Added options.html page. I will write this later.

* Word Selection Logic and others

Content_script.js
1. Added the event listener that enables the detection of selected word(s). 

manifest.json
1. Solved a bug of an extra comma

popup.html
1. A dummy html for the structure to work. 

background.js
1. Added for the console.log for me to know that it has been loaded.

* Created wordLookup function

Word lookup function will return the link to Cambridge dictionary if vocab is a single word. If vocab has any spaces in it then it will replace it with +. 

To do for the future. 
1. Find way to remove the starting space and ending space for both single vocab or string.

* Added a few functions

1. Create a hovering "speech" bubble over text if it is selected. 
2. Click anywhere else to remove it if the speech bubble is available. 
3. Removed the regular expression and the Cambridge link. I am going to switch to using Google Dictionary API by meetDeveloper and Oxford Dictionary API.

* Append Definition to Div

I am able to call the API and then append the definition to the webpage. 

Further to do. 
1. Handle when there are no definitions returned. 
2. How to handle multiple definitions.

* Added a few function

1. Create a hovering "speech" bubble over text if it is selected. 
2. Click anywhere else to remove it if the speech bubble is available. 
3. Removed the regular expression and the Cambridge link. I am going to switch to using Google Dictionary API by meetDeveloper and Oxford Dictionary API.

* Append Definition to Div

I am able to call the API and then append the definition to the webpage. 

Further to do. 
1. Handle when there are no definitions returned. 
2. How to handle multiple definitions.

* Small Modification

1. Added selectedText into createDiv. This is to make the "more" button usable and can link to google Dictionary. 
2. Changed the val.forEach to access the first definition. This will help the definition to be clearer.

* Refactoring and Displaying Phonetics

1. Refactored the styling of popup for better readability. 
2. Added the capability to display phonetics and part of speech in the pop up. 
3. Added the logic to get the first instance of explanation.
4. Added logic to handle the exception when there is no definition in the head of array.

* Click Enhancement, No Definition Handling and removed console.logs

1. Click enhancement
- Changed from onmouseup to onclick. 
- This enables the "more" link to be clicked. 

2. No definition handling
- This enables user to know that our program can't find the definition of the highlighted word

3. console.log removals
- This ensures the program is easier to read and doesn't flood the console with unnecessary information.